requirements.txt
aiohttp==3.8.0
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
atpublic==2.3
attrs==21.2.0
black==21.10b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cycler==0.11.0
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.6.0
dpath==2.0.5
dulwich==0.20.26
dvc==2.8.3
flatten-dict==0.4.2
flufl.lock==6.0
frozenlist==1.2.0
fsspec==2021.11.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
grandalf==0.6
idna==3.3
iniconfig==1.1.1
kiwisolver==1.3.2
mailchecker==4.0.16
matplotlib==3.4.3
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
nanotime==0.5.2
networkx==2.6.3
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pathspec==0.9.0
phonenumbers==8.12.37
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
ply==3.11
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydot==1.4.2
pygame==2.1.0
pygit2==1.7.0
Pygments==2.10.0
pygtrie==2.4.2
pyparsing==2.4.7
pytest==6.2.5
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
rich==10.13.0
rope==0.21.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.7.2
shortuuid==1.0.8
shtab==1.4.2
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cu113
tqdm==4.62.3
types-PyYAML==6.0.1
typing-extensions==3.10.0.2
urllib3==1.26.7
voluptuous==0.12.2
wcwidth==0.2.5
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
